Support string references to classes in descriptors
Entities needed until now to be defined in a particular order in
entities.py in order to reference other entities. Because of this
limitation (I assume), parts of the metadata have been pushed to the
bottom of the file:

$ git show 0.3.9:genologics/entities.py | tail -n 5

This fix allows descriptors to refer to entities with a string
rather than having to use the type directly, which solves this issue.